% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{Helper function that splits vector into chunks of specific size}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{The vector to be chunked.}

\item{n}{The prefered size of those chunks.}
}
\description{
Credit goes to: Harlan (http://stackoverflow.com/users/135944/harlan) and
dfrankow (http://stackoverflow.com/users/34935/dfrankow) due to their efforts
at Stackoverflow (http://stackoverflow.com/a/3321659/1144966)
}

