% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract_access_counts}
\alias{extract_access_counts}
\title{Helper function for wp_trend}
\usage{
extract_access_counts(wikipediaTrendJson)
}
\arguments{
\item{wikipediaTrendJson}{The results from server request in JSON format}
}
\description{
The function does accepts a list of JSON snippets and transforms the information into a data frame
}

