% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pversion_inc.r
\docType{data}
\name{files}
\alias{files}
\title{function that increases the packages version
modified version of ...
## by:     alko989
## source: http://stackoverflow.com/q/24209336/1144966}
\format{\preformatted{ chr [1:2] "d:/peter/dropbox/rpackages/wikipediatrend/DESCRIPTION" ...
}}
\usage{
files
}
\description{
function that increases the packages version
modified version of ...
## by:     alko989
## source: http://stackoverflow.com/q/24209336/1144966
}
\keyword{datasets}

