% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html}
\alias{html}
\title{function downloading content}
\usage{
html(url, ...)
}
\arguments{
\item{url}{url to content}

\item{...}{further arguments passed trough to httr::GET()}
}
\description{
function downloading content
}
