% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_plot.R
\name{plot.wp_df}
\alias{plot.wp_df}
\title{simple plot function for objects of class wp_df}
\usage{
\method{plot}{wp_df}(wp_df, ...)
}
\arguments{
\item{wp_df}{object of class wp_df to be plotted}
}
\description{
simple plot function for objects of class wp_df
}

