% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_df.R
\name{print.wp_df}
\alias{print.wp_df}
\title{a wikipediastrend specific data frame}
\usage{
\method{print}{wp_df}(x, ...)
}
\arguments{
\item{x}{the thing to be printed}

\item{...}{print.default parameters}
}
\description{
a wikipediastrend specific data frame
}

