% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa_pversion_inc.r
\name{pversion_inc}
\alias{pversion_inc}
\title{function that increases the packages version
modified version of ...
## by:     alko989
## source: http://stackoverflow.com/q/24209336/1144966}
\usage{
pversion_inc()
}
\description{
function that increases the packages version
modified version of ...
## by:     alko989
## source: http://stackoverflow.com/q/24209336/1144966
}

