% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_utf8_csv.R
\name{read_utf8_csv}
\alias{read_utf8_csv}
\title{function to read csv file with UTF-8 characters (even under Windwos) that
were created by write_U}
\usage{
read_utf8_csv(file)
}
\arguments{
\item{file}{file name / path where to put the data}

\item{df}{data frame to be written to file}
}
\description{
function to read csv file with UTF-8 characters (even under Windwos) that
were created by write_U
}

