% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_utf8_csv.R
\name{toUTF8}
\alias{toUTF8}
\title{function to convert character vectors to UTF-8 encoding}
\usage{
toUTF8(x)
}
\arguments{
\item{x}{the vector to be converted}
}
\description{
function to convert character vectors to UTF-8 encoding
}

