% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_search.R
\name{wiki_search}
\alias{wiki_search}
\title{wiki_search}
\usage{
wiki_search(page_fragment = "R_(programming_language)", lang = "en")
}
\arguments{
\item{page_fragment}{substring to search for}

\item{lang}{wikipedia language to search in}
}
\description{
wiki_search
}
