% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_cache.R
\name{wp_add_to_cache}
\alias{wp_add_to_cache}
\title{function adding downloaded data to the cache}
\usage{
wp_add_to_cache(df)
}
\arguments{
\item{df}{data frame to be aded to cache}
}
\description{
function adding downloaded data to the cache
}
