% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_cache.R
\name{wp_cache_file}
\alias{wp_cache_file}
\title{function for determining where downloaded information can be stored}
\usage{
wp_cache_file()
}
\description{
function for determining where downloaded information can be stored
}
