% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_cache.R
\name{wp_cache_reset}
\alias{wp_cache_reset}
\title{function for resetting cache/cachefile as if package was just loaded}
\usage{
wp_cache_reset()
}
\description{
function for resetting cache/cachefile as if package was just loaded
}

