% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wp_check_date_inputs}
\alias{wp_check_date_inputs}
\title{Helper function for \code{wp_trend()}}
\usage{
wp_check_date_inputs(from, to)
}
\arguments{
\item{from}{first date of timespan to check}

\item{to}{second date of timespan to check}
}
\description{
Function checks time span given by fromDate and toData if complying with
logical constraints: from not prior to first available data; to not past
today
}

