% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wp_day}
\alias{wp_day}
\alias{wp_day.POSIXlt}
\alias{wp_day.default}
\title{Function to extract the day}
\usage{
wp_day(timestamp)

\method{wp_day}{POSIXlt}(timestamp)

\method{wp_day}{default}(timestamp)
}
\arguments{
\item{timestamp}{a timestamp}
}
\description{
Function to extract the day from a timestamp of e.g. class POSIXlt, POSIXct, Date, or character. If the option is of type character it should be in the form of yyyy-mm-dd.
}
\section{Methods (by class)}{
\itemize{
\item \code{POSIXlt}: extract

\item \code{default}: extract
}}

