% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wp_expand_ts}
\alias{wp_expand_ts}
\title{Helper function for wp_trend that transforms a time span into a series of months}
\usage{
wp_expand_ts(from, to, by)
}
\arguments{
\item{from}{start of timespan}

\item{to}{end of timespan}
}
\description{
Helper function for wp_trend that transforms a time span into a series of months
}

