% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wp_extract_data}
\alias{wp_extract_data}
\title{Helper function for wp_trend}
\usage{
wp_extract_data(wp_json)
}
\arguments{
\item{wikipediaTrendJson}{The results from server request in JSON format}
}
\description{
The function does accepts a list of JSON snippets and transforms the information into a data frame
}

