% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wp_friendly_save}
\alias{wp_friendly_save}
\title{Helper function for wp_trend}
\usage{
wp_friendly_save(res, friendly, resname)
}
\arguments{
\item{res}{internal data retreived by \code{wp_trend()}}

\item{friendly}{should retrieved data be saved for later use (option is
passed down from \code{wp_trend()})}

\item{resname}{the name serving to save the data (option is passed down from
\code{wp_trend()})}
}
\description{
Function writes wp_trend data to a CSV file that can be used for subsequent
requests
}

