% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wp_http_header.R
\name{wp_http_header}
\alias{wp_http_header}
\title{function for pasting together the HTTP user-agent field}
\usage{
wp_http_header()
}
\description{
function for pasting together the HTTP user-agent field
}

