% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wp_join_data.R
\name{wp_join_data}
\alias{wp_join_data}
\title{function for joining cache and request data}
\usage{
wp_join_data(request, cache)
}
\arguments{
\item{request}{data downloaded}

\item{cache}{data loaded from cache file}
}
\description{
function for joining cache and request data
}

