% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wp_jsons_to_df.R
\name{wp_jsons_to_df}
\alias{wp_jsons_to_df}
\title{Helper function for wp_trend}
\usage{
wp_jsons_to_df(wp_json)
}
\arguments{
\item{wp_json}{The results from server request in JSON format}
}
\description{
The function does accepts a list of JSON snippets and transforms the
information into a data frame
}

