% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wp_linked_pages.R
\name{wp_linked_pages}
\alias{wp_linked_pages}
\title{function looking for other languages of a page}
\usage{
wp_linked_pages(page, lang)
}
\arguments{
\item{page}{title of the wikipedia article to look for links to other
languages}

\item{lang}{language (shorthand, e.g. "en" for English or "de" for German and
so on) for the afor given title}
}
\description{
function looking for other languages of a page
}

