% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_load_and_save.r
\name{wp_load}
\alias{wp_load}
\title{Helper function for wp_trend that loads previous saved wp_trend results}
\usage{
wp_load(file = wp_cache_file())
}
\arguments{
\item{file}{name of the file from which previous \code{wp_trend()}
results should be loaded}
}
\description{
Helper function for wp_trend that loads previous saved wp_trend results
}

