% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wp_month}
\alias{wp_month}
\alias{wp_month.POSIXlt}
\alias{wp_month.default}
\title{Function to extract the month}
\usage{
wp_month(timeStamp)

\method{wp_month}{POSIXlt}(timeStamp)

\method{wp_month}{default}(timeStamp)
}
\arguments{
\item{timestamp}{a timestamp}
}
\description{
Function to extract the month from a timestamp of e.g. class POSIXlt, POSIXct, Date, or character. If the option is of type character it should be in the form of yyyy-mm-dd.
}
\section{Methods (by class)}{
\itemize{
\item \code{POSIXlt}: extract

\item \code{default}: extract
}}

