% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_prepare_date_for_pageviews.R
\name{wp_prepare_date_for_pageviews}
\alias{wp_prepare_date_for_pageviews}
\title{function to prepare date for execution of pageviews request}
\usage{
wp_prepare_date_for_pageviews(date, type)
}
\arguments{
\item{date}{date as character string}

\item{type}{"from" or "to"}
}
\description{
function to prepare date for execution of pageviews request
}
\keyword{internal}
