% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wp_save.R
\name{wp_save}
\alias{wp_save}
\title{Helper function for wp_trend}
\usage{
wp_save(res, file)
}
\arguments{
\item{res}{internal data retreived by \code{wp_trend()}}

\item{friendly}{should retrieved data be saved for later use (option is
passed down from \code{wp_trend()})}

\item{resname}{the name serving to save the data (option is passed down from
  \code{wp_trend()})}
}
\description{
Function writes wp_trend data to a CSV file that can be used for subsequent
requests
}

