% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wp_cache.R
\name{wp_set_cache_file}
\alias{wp_set_cache_file}
\title{function to set a different cachefile as session default}
\usage{
wp_set_cache_file(file)
}
\arguments{
\item{file}{name of the file to be used as cache}
}
\description{
function to set a different cachefile as session default
}

