% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_df.R
\name{wp_set_print_options}
\alias{wp_set_print_options}
\title{function for setting print options for print.wp_df()}
\usage{
wp_set_print_options(x)
}
\arguments{
\item{x}{a list of options, e.g. list(rows=35, width=50) or list(rows=Inf, width=Inf)}
}
\description{
function for setting print options for print.wp_df()
}

