% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wp_trend}
\alias{wp_trend}
\title{Function for getting access statistics for wikipedia pages}
\usage{
wp_trend(requestFrom = "retep.meissner@gmail.com",
  pageName = "Peter_Principle", countryCode = "en", fromDate = Sys.Date()
  - 30, toDate = Sys.Date(), friendly = T)
}
\arguments{
\item{requestFrom}{This parameter sends an identifier (i.e. your email
  address to contact you if necessary) to stats.grok.se (the server that
  kindly provides the information you request).}

\item{pageName}{The name of the Wikipedia page as to be found in the URL to the wikipedia article. If e.g. the URL is: \code{https://en.wikipedia.org/wiki/Peter_Fox_(musician)}, than the page name equals to \code{Peter_Fox_(musician)}.}

\item{countryCode}{The country code of}

\item{fromDate}{blsa}

\item{toDate}{blsa}

\item{friendly}{blsa}
}
\description{
Function for getting access statistics for wikipedia pages
}

