% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wp_yearmonth.R
\name{wp_yearmonth}
\alias{wp_yearmonth}
\title{function for getting year and month of a timestamp}
\usage{
wp_yearmonth(timestamp)
}
\arguments{
\item{timestamp}{Some sort of timestamp e.g. of class POSIXlt, POSIXct, Date,
or character. If the option is of type character it should be in the form
of yyyy-mm-dd.}
}
\description{
function for getting year and month of a timestamp
}

