% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_decompress.R
\name{wpd_decompress}
\alias{wpd_decompress}
\title{wpd_decompress}
\usage{
wpd_decompress(cont_item, lang)
}
\arguments{
\item{cont_item}{single content item}

\item{lang}{language to decompress}
}
\description{
wpd_decompress
}
