% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_search.R
\name{wpd_search}
\alias{wpd_search}
\title{wpd_search}
\usage{
wpd_search(page_fragment = "R_(programming_language)", lang = "en")
}
\arguments{
\item{page_fragment}{substring to search for within titles}

\item{lang}{language to search}
}
\description{
wpd_search
}
