% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_utf8_csv.R
\name{write_utf8_csv}
\alias{write_utf8_csv}
\title{function to write csv files with UTF-8 characters (even under Windwos)}
\usage{
write_utf8_csv(df, file)
}
\arguments{
\item{df}{data frame to be written to file}

\item{file}{file name / path where to put the data
#}
}
\description{
function to write csv files with UTF-8 characters (even under Windwos)
}

