% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wp_cache.R
\docType{data}
\name{cache}
\alias{cache}
\title{a cache for downloads}
\format{\preformatted{<environment: 0x0000000008982b78> 
}}
\usage{
cache
}
\description{
a cache for downloads
}
\keyword{datasets}

