% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wp_df.R
\name{print.wp_df}
\alias{print}
\alias{print.wp_df}
\title{a wikipediastrend specific data frame}
\usage{
\method{print}{wp_df}(df, ..., digits = NULL, quote = FALSE, right = TRUE,
  row.names = TRUE)
}
\description{
a wikipediastrend specific data frame
}

