% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wp_load_and_save.r
\name{read_utf8_csv}
\alias{read_utf8_csv}
\title{function to read csv file with UTF-8 characters (even under Windwos) that
were created by write_U}
\usage{
read_utf8_csv(file)
}
\arguments{
\item{file}{file name / path where to get the data}
}
\description{
function to read csv file with UTF-8 characters (even under Windwos) that
were created by write_U
}

