% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_load_and_save.r
\name{toUTF8}
\alias{toUTF8}
\title{function to convert character vectors to UTF-8 encoding}
\usage{
toUTF8(x)
}
\arguments{
\item{x}{the vector to be converted}
}
\description{
function to convert character vectors to UTF-8 encoding
}

