% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_cache.R
\docType{data}
\name{wp_cache}
\alias{wp_cache}
\title{a cache for downloads}
\format{An object of class \code{environment} of length 2.}
\usage{
wp_cache
}
\description{
a cache for downloads
}
\keyword{datasets}
