% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wp_cache.R
\docType{data}
\name{wp_cache}
\alias{wp_cache}
\title{a cache for downloads}
\format{\preformatted{<environment: 0x00000000092033b8> 
}}
\usage{
wp_cache
}
\description{
a cache for downloads
}
\keyword{datasets}

