% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wp_cache_file.R
\name{wp_cache_file}
\alias{wp_cache_file}
\title{function for determining where downloaded information can be stored}
\usage{
wp_cache_file()
}
\description{
function for determining where downloaded information can be stored
}

