% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_check_date_inputs.R
\name{wp_check_date_inputs}
\alias{wp_check_date_inputs}
\title{Helper function for \code{wp_trend()}}
\usage{
wp_check_date_inputs(from, to)
}
\arguments{
\item{from}{first date of timespan to check}

\item{to}{second date of timespan to check}
}
\description{
Function that checks if the time span given by from and to (passed down from
wp_trend) are complying with logical constraints: from not prior to first
available data; to not past today; to not prior to from
}
