% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wp_day.R
\name{wp_day}
\alias{wp_day}
\alias{wp_day.POSIXlt}
\alias{wp_day.default}
\title{Function to extract the day}
\usage{
wp_day(timestamp)

\method{wp_day}{POSIXlt}(timestamp)

\method{wp_day}{default}(timestamp)
}
\arguments{
\item{timestamp}{Some sort of timestamp e.g. of class POSIXlt, POSIXct, Date,
  or character. If the option is of type character it should be in the form
  of yyyy-mm-dd.}
}
\description{
Function to extract the day from a timestamp of e.g. class POSIXlt, POSIXct,
Date, or character. If the option is of type character it should be in the
form of yyyy-mm-dd.
}
\section{Methods (by class)}{
\itemize{
\item \code{POSIXlt}: extract day from timestamp

\item \code{default}: extract day from timestamp
}}

