% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_download_data.R
\name{wp_download_data}
\alias{wp_download_data}
\title{function downloading prepared URLs}
\usage{
wp_download_data(urls, wait = 1)
}
\arguments{
\item{urls}{a vector of urls to be downloaded}

\item{wait}{the time to wait in seconds before downloading the next chunk (default=1)}
}
\description{
function downloading prepared URLs
}

