% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wp_expand_ts.R
\name{wp_expand_ts}
\alias{wp_expand_ts}
\title{Helper function for wp_trend that transforms a time span into a series of
months}
\usage{
wp_expand_ts(from, to, by)
}
\arguments{
\item{from}{start of timespan}

\item{to}{end of timespan}

\item{by}{in which time unit expansion should take place: \code{"day"},
  \code{"month"}, \code{"year"}}
}
\description{
Helper function for wp_trend that transforms a time span into a series of
months
}

