% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wp_friendly_load}
\alias{wp_friendly_load}
\title{Helper function for wp_trend that loads previous saved wp_trend results}
\usage{
wp_friendly_load(resname, friendly)
}
\description{
Helper function for wp_trend that loads previous saved wp_trend results
}

