% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wp_friendly_load}
\alias{wp_friendly_load}
\title{Helper function for wp_trend that loads previous saved wp_trend results}
\usage{
wp_friendly_load(resname, friendly)
}
\arguments{
\item{resname}{name of the file from which previous \code{wp_trend()} results should be loaded (option is passed down from \code{wp_trend()})}

\item{friendly}{should previous results be ignored or loaded (option is passed down from \code{wp_trend()})}
}
\description{
Helper function for wp_trend that loads previous saved wp_trend results
}

