% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_get_data.R
\name{wp_get_data}
\alias{wp_get_data}
\title{function for getting data (download + extraction)}
\usage{
wp_get_data(urls)
}
\arguments{
\item{urls}{urls to be downloaded}
}
\description{
function for getting data (download + extraction)
}
