% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_get_data.R
\name{wp_get_data}
\alias{wp_get_data}
\title{function for getting data (download + extraction)}
\usage{
wp_get_data(page = "R_(programming_language)", from = "2007-12-01",
  to = as.character(Sys.Date()), lang = "en", user_type = "all",
  platform = "all")
}
\arguments{
\item{urls}{urls to be downloaded}
}
\description{
function for getting data (download + extraction)
}
