% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wp_prepare_urls.R
\name{wp_prepare_urls}
\alias{wp_prepare_urls}
\title{function preparing URLs for downlaod}
\usage{
wp_prepare_urls(page, from, to, lang)
}
\arguments{
\item{page}{page title supplied by wp_trend()}

\item{from}{from supplied by wp_trend()}

\item{to}{to supplied by wp_trend()}

\item{lang}{lang supplied by wp_trend()}
}
\description{
function preparing URLs for downlaod
}

