% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wp_save.R
\name{wp_save}
\alias{wp_save}
\title{Helper function for wp_trend}
\usage{
wp_save(res, file = wp_cache_file())
}
\arguments{
\item{res}{data retreived by \code{wp_trend()}}

\item{file}{where to save/cache results (defaults to \code{wp_cache_file()})}
}
\description{
Function writes wp_trend() data to a CSV file that serves as cache/storage
}

