% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wp_trend}
\alias{wp_trend}
\title{Function for getting access statistics for wikipedia pages}
\usage{
wp_trend(requestFrom = "anonymous", pageName = "Peter_principle",
  countryCode = "en", fromDate = Sys.Date() - 30, toDate = Sys.Date(),
  friendly = F)
}
\arguments{
\item{requestFrom}{This parameter sends an identifier (i.e. your email
  address to contact you if necessary) to stats.grok.se (the server that
  kindly provides the information you request).}

\item{pageName}{The name of the Wikipedia page as to be found in the URL to the wikipedia article. If e.g. the URL is: \code{https://en.wikipedia.org/wiki/Peter_Fox_(musician)}, than the page name equals to \code{Peter_Fox_(musician)}.}

\item{countryCode}{The country code of}

\item{fromDate}{blsa}

\item{toDate}{blsa}

\item{friendly}{Either \code{TRUE}, \code{FALSE}, \code{1} or \code{2}
This option causes twofold. First, the results of the request are saved in the current working directory in a CSV file with name scheme: \code{wp__[page name]_[country code].csv}.
Second, the function will look if perhaps a previously saved result is available to be used to only download those information that are still missing instead of the whole timespan.

For storage on disk write.csv() and write.csv2() are used}
}
\description{
Function for getting access statistics for wikipedia pages
}

