% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_trend.R
\name{wp_trend}
\alias{wp_trend}
\title{Function for getting access statistics for wikipedia pages}
\usage{
wp_trend(page, from = prev_month_start(), to = prev_month_end(),
  lang = "en", file = "", friendly, requestFrom, userAgent)
}
\arguments{
\item{page}{The name of the Wikipedia page as to be found in the URL to the 
wikipedia article. If e.g. the URL is: 
\code{https://en.wikipedia.org/wiki/Peter_Fox_(musician)}, than the page 
name equals to \code{Peter_Fox_(musician)}.}

\item{from}{The starting date of the timespan for which access statistics 
should be retrieved - note that there is no data prior to 2007-12-01. 
Supply some sort of timestamp e.g. of class POSIXlt, POSIXct, Date, or 
character. If the option is of type character it should be in the form of 
yyyy-mm-dd.}

\item{to}{The last date for which access statistics should be retrieved. 
Supply some sort of timestamp e.g. of class POSIXlt, POSIXct, Date, or 
character. If the option is of type character it should be in the form of 
yyyy-mm-dd.}

\item{lang}{The language shorthand identifying which Wikipedia access 
statistics are to be used: e.g. \code{"en"} for the English version found 
at https://en.wikipedia.org, \code{"de"} for the German version found at 
https://de.wikipedia.org or perhaps \code{"als"} for the Alemannic dialect 
found under https://als.wikipedia.org/.}

\item{file}{Where to store data retrieved by function call? The data 
is stored in CSV format. If an already existing file is used for storage, 
the old data will not be deleted but instead new data will be added to this 
file! The cache in memory before the function call will be saved and restored 
afterwards. If you prefer automatic mirroring of in memory cache and cache 
on disk}

\item{friendly}{deprecated}

\item{requestFrom}{deprecated}

\item{userAgent}{deprecated}
}
\description{
Function for getting access statistics for wikipedia pages
}
\examples{
library(wikipediatrend)
wp_trend(page        = c("Cheese", "K\\u00e4se"),
         from        = "2014-11-01", 
         to          = "2014-11-30", 
         lang        = c("en", "de"),
         file        = wp_cache_file()
         )


}

