% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wp_date_functions.r
\name{wp_wday}
\alias{wp_wday}
\alias{wp_wday.POSIXlt}
\alias{wp_wday.default}
\title{Function to extract the day of the week}
\usage{
wp_wday(timestamp, startmonday = T)

\method{wp_wday}{POSIXlt}(timestamp, startmonday = T)

\method{wp_wday}{default}(timestamp, startmonday = T)
}
\arguments{
\item{timestamp}{Some sort of timestamp e.g. of class POSIXlt, POSIXct, Date,
  or character. If the option is of type character it should be in the form
  of yyyy-mm-dd.}

\item{startmonday}{whether the week should start on Monday (TRUE, the
  default) or it should start on Sunday (FALSE)}
}
\description{
Function to extract the day from a timestamp of e.g. class POSIXlt, POSIXct,
Date, or character. If the option is of type character it should be in the
form of yyyy-mm-dd.
}
\section{Methods (by class)}{
\itemize{
\item \code{POSIXlt}: 

\item \code{default}: 
}}

