% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wp_year}
\alias{wp_year}
\alias{wp_year.POSIXlt}
\alias{wp_year.default}
\title{Function to extract the year}
\usage{
wp_year(timestamp)

\method{wp_year}{POSIXlt}(timestamp)

\method{wp_year}{default}(timestamp)
}
\arguments{
\item{timestamp}{a timestamp}
}
\description{
Function to extract the year from a timestamp of e.g. class POSIXlt, POSIXct, Date, or character. If the option is of type character it should be in the form of yyyy-mm-dd.
}
\section{Methods (by class)}{
\itemize{
\item \code{POSIXlt}: extract

\item \code{default}: extract
}}

